// src/apis/models/Volo_Saas_Host_Dtos_EditionCreateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Saas_Host_Dtos_EditionCreateDto = {
    readonly extraProperties?: Record<string, any> | null;
    displayName: string;
    planId?: string | null;
};